Support get command name in macos
